requirements.txt
langchain==0.2.14
langchain-community==0.2.12
langchain-groq===0.1.9
pypdf==5.0.0
streamlit==1.35.0
python-dotenv==1.0.1
beautifulsoup4==4.12.3
